+++ composeApp/src/androidMain/kotlin/org/onion/agent/MainActivity.kt
@@ -8,12 +8,14 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import io.github.vinceglb.filekit.FileKit
 import io.github.vinceglb.filekit.dialogs.init
+import org.onion.agent.database.initAgentDatabase
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         enableEdgeToEdge()
         super.onCreate(savedInstanceState)
         FileKit.init(this)
+        initAgentDatabase(this)
         setContent {
             App()
         }